CORE-5386 Fixed UUID conversions in schema rewrite.
Some IDs in some tables are not valid UUIDs.
Added a function that converts an id column for a given table to UUID, generating a new UUID if the current value is invalid.
Updated the conversions for tools, tasks, apps, file_parameters, parameters, parameter_groups, and validation_rules to use this function.
Updated the conversions for parameter_types and rule_type to remove invalid prefixes from their id values when converting to UUIDs.
Updated the conversion for workflow_io_maps to delete input and output mappings that no longer refer to existing file_parameters, which removes some invalid UUID values in these colums.