Fixed a bunch of interapps related issues.
There's now a foreign key to the interactive_apps_proxy_settings table
in the container_settings table.

The foreign key into the container_settings table was dropped from the
interactive_apps_proxy_settings table.

A uniqueness constraint was added to the
interactive_apps_proxy_settings.id column.

The SQL files for setting up the various tables and constraints were
modified to account for the above.

A new version entry was added to the database.